chore: remove componentsX updates
